requirements.txt
# === Core ML / Embeddings ===
sentence-transformers>=2.2.2
torch>=2.0.0

# === File type detection ===
python-magic-bin; platform_system == "Windows"
python-magic; platform_system != "Windows"

# === Document parsing ===
python-docx>=1.1.0
pypdf>=4.0.0

# === Image metadata (EXIF, future CLIP integration) ===
Pillow>=10.0.0

# === Utilities ===
numpy>=1.24
tqdm>=4.66